+++ dashboard/client/src/services/data-providers/apps-list.ts
@@ -19,7 +19,7 @@ import { AppsListRequest } from '../../pb/dashboard_pb';
 import * as PbDashboard from '../../pb/dashboard_pb';
 
 const client = new DashboardClient(
-  process.env.REACT_APP_DASHBOARD_GRPC_WEB_URL || 'http://localhost:8080/api'
+  process.env.REACT_APP_DASHBOARD_GRPC_WEB_URL || 'http://localhost:8081/api'
 );
 
 export const getAppsList = async (): Promise<PbDashboard.AppsListResponse.AsObject | null> => {
@@ -35,7 +35,7 @@ export const getAppsList = async (): Promise<PbDashboard.AppsListResponse.AsObje
           if (err || !response) {
             reject(err);
           }
-          resolve(response?.toObject());
+          resolve(response ?.toObject());
         }
       );
     }