+++ launcher/cli/apps.go
@@ -28,6 +28,7 @@ import (
 	"github.com/dfuse-io/bstream"
 	_ "github.com/dfuse-io/dauth/null" // register plugin
 	abicodecApp "github.com/dfuse-io/dfuse-eosio/abicodec/app/abicodec"
+	"github.com/dfuse-io/dfuse-eosio/apiproxy"
 	dblockmeta "github.com/dfuse-io/dfuse-eosio/blockmeta"
 	"github.com/dfuse-io/dfuse-eosio/codec"
 	"github.com/dfuse-io/dfuse-eosio/dashboard"
@@ -58,13 +59,12 @@ import (
 )
 
 func init() {
-	// Manager
 	launcher.RegisterApp(&launcher.AppDef{
 		ID:          "manager",
 		Title:       "Producer node",
 		Description: "Block producing node",
 		MetricsID:   "manager",
-		Logger:      newLoggerDef("github.com/dfuse-io/manageos/app/nodeos_manager", []zapcore.Level{zap.WarnLevel, zap.WarnLevel, zap.InfoLevel, zap.DebugLevel}),
+		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/manageos/app/nodeos_manager", []zapcore.Level{zap.WarnLevel, zap.WarnLevel, zap.InfoLevel, zap.DebugLevel}),
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().String("manager-api-addr", EosManagerAPIAddr, "eos-manager API address")
 			cmd.Flags().String("manager-nodeos-api-addr", NodeosAPIAddr, "Target API address")
@@ -154,16 +154,15 @@ func init() {
 		},
 	})
 
-	// Mindreader
 	launcher.RegisterApp(&launcher.AppDef{
 		ID:          "mindreader",
 		Title:       "Reader node",
 		Description: "Blocks reading node",
 		MetricsID:   "manager",
-		Logger:      newLoggerDef("github.com/dfuse-io/manageos/(app/nodeos_mindreader|mindreader).*", []zapcore.Level{zap.WarnLevel, zap.WarnLevel, zap.InfoLevel, zap.DebugLevel}),
+		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/manageos/(app/nodeos_mindreader|mindreader).*", []zapcore.Level{zap.WarnLevel, zap.WarnLevel, zap.InfoLevel, zap.DebugLevel}),
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().String("mindreader-manager-api-addr", EosMindreaderHTTPAddr, "eos-manager API address")
-			cmd.Flags().String("mindreader-api-addr", NodeosAPIAddr, "Target API address")
+			cmd.Flags().String("mindreader-nodeos-api-addr", NodeosAPIAddr, "Target API address")
 			cmd.Flags().Bool("mindreader-connection-watchdog", false, "Force-reconnect dead peers automatically")
 			cmd.Flags().String("mindreader-config-dir", "mindreadernode/config", "Directory for config files. ")
 			cmd.Flags().String("mindreader-nodeos-path", NodeosBinPath, "Path to the nodeos binary. Defaults to the nodeos found in your PATH")
@@ -252,7 +251,7 @@ to find how to install it.`)
 
 			return nodeosMindreaderApp.New(&nodeosMindreaderApp.Config{
 				ManagerAPIAddress:          viper.GetString("mindreader-manager-api-addr"),
-				NodeosAPIAddress:           viper.GetString("mindreader-api-addr"),
+				NodeosAPIAddress:           viper.GetString("mindreader-nodeos-api-addr"),
 				ConnectionWatchdog:         viper.GetBool("mindreader-connection-watchdog"),
 				NodeosConfigDir:            buildStoreURL(viper.GetString("global-data-dir"), viper.GetString("mindreader-config-dir")),
 				NodeosBinPath:              viper.GetString("mindreader-nodeos-path"),
@@ -293,7 +292,7 @@ to find how to install it.`)
 		Title:       "Relayer",
 		Description: "Serves blocks as a stream, with a buffer",
 		MetricsID:   "relayer",
-		Logger:      newLoggerDef("github.com/dfuse-io/relayer.*", nil),
+		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/relayer.*", nil),
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().String("relayer-grpc-listen-addr", RelayerServingAddr, "Address to listen for incoming gRPC requests")
 			cmd.Flags().StringSlice("relayer-source", []string{MindreaderGRPCAddr}, "List of blockstream sources to connect to for live block feeds (repeat flag as needed)")
@@ -321,13 +320,12 @@ to find how to install it.`)
 		},
 	})
 
-	// Merger
 	launcher.RegisterApp(&launcher.AppDef{
 		ID:          "merger",
 		Title:       "Merger",
 		Description: "Produces merged block files from single-block files",
 		MetricsID:   "merger",
-		Logger:      newLoggerDef("github.com/dfuse-io/merger.*", nil),
+		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/merger.*", nil),
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().String("merger-merged-block-path", MergedBlocksFilesPath, "URL of storage to write merged-block-files to")
 			cmd.Flags().String("merger-one-block-path", OneBlockFilesPath, "URL of storage to read one-block-files from")
@@ -389,13 +387,12 @@ to find how to install it.`)
 		},
 	})
 
-	// fluxdb
 	launcher.RegisterApp(&launcher.AppDef{
 		ID:          "fluxdb",
 		Title:       "FluxDB",
 		Description: "Temporal chain state store",
 		MetricsID:   "fluxdb",
-		Logger:      newLoggerDef("github.com/dfuse-io/dfuse-eosio/fluxdb.*", nil),
+		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/dfuse-eosio/fluxdb.*", nil),
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().Bool("fluxdb-enable-server-mode", true, "Enables flux server mode, launch a server")
 			cmd.Flags().Bool("fluxdb-enable-inject-mode", true, "Enables flux inject mode, writes into KVDB")
@@ -430,13 +427,12 @@ to find how to install it.`)
 		},
 	})
 
-	// KVDB Loader
 	launcher.RegisterApp(&launcher.AppDef{
 		ID:          "kvdb-loader",
 		Title:       "DB loader",
 		Description: "Main blocks and transactions database",
 		MetricsID:   "kvdb-loader",
-		Logger:      newLoggerDef("github.com/dfuse-io/dfuse-eosio/kvdb-loader.*", nil),
+		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/dfuse-eosio/kvdb-loader.*", nil),
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().String("kvdb-loader-chain-id", "", "Chain ID")
 			cmd.Flags().String("kvdb-loader-processing-type", "live", "The actual processing type to perform, either `live`, `batch` or `patch`")
@@ -483,7 +479,7 @@ to find how to install it.`)
 		Title:       "Blockmeta",
 		Description: "Serves information about blocks",
 		MetricsID:   "blockmeta",
-		Logger:      newLoggerDef("github.com/dfuse-io/blockmeta.*", nil),
+		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/blockmeta.*", nil),
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().String("blockmeta-grpc-listen-addr", BlockmetaServingAddr, "Address to listen for incoming gRPC requests")
 			cmd.Flags().String("blockmeta-block-stream-addr", RelayerServingAddr, "Websocket endpoint to get a real-time blocks feed")
@@ -528,7 +524,7 @@ to find how to install it.`)
 		Title:       "ABI codec",
 		Description: "Decodes binary data against ABIs for different contracts",
 		MetricsID:   "abicodec",
-		Logger:      newLoggerDef("github.com/dfuse-io/dfuse-eosio/abicodec.*", nil),
+		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/dfuse-eosio/abicodec.*", nil),
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().String("abicodec-grpc-listen-addr", AbiServingAddr, "Address to listen for incoming gRPC requests")
 			cmd.Flags().String("abicodec-search-addr", RouterServingAddr, "Base URL for search service")
@@ -563,7 +559,7 @@ to find how to install it.`)
 		Title:       "Search indexer",
 		Description: "Indexes transactions for search",
 		MetricsID:   "search-indexer",
-		Logger:      newLoggerDef("github.com/dfuse-io/search/(indexer|app/indexer).*", nil),
+		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/search/(indexer|app/indexer).*", nil),
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().String("search-indexer-grpc-listen-addr", IndexerServingAddr, "Address to listen for incoming gRPC requests")
 			cmd.Flags().String("search-indexer-http-listen-addr", IndexerHTTPServingAddr, "Address to listen for incoming http requests")
@@ -622,7 +618,7 @@ to find how to install it.`)
 		Title:       "Search router",
 		Description: "Routes search queries to archiver, live",
 		MetricsID:   "search-router",
-		Logger:      newLoggerDef("github.com/dfuse-io/search/(router|app/router).*", nil),
+		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/search/(router|app/router).*", nil),
 		RegisterFlags: func(cmd *cobra.Command) error {
 			// Register common search flags once for all the services
 			cmd.Flags().String("search-common-mesh-store-addr", "", "[COMMON] Address of the backing etcd cluster for mesh service discovery.")
@@ -630,7 +626,7 @@ to find how to install it.`)
 			cmd.Flags().String("search-common-mesh-service-version", DmeshServiceVersion, "[COMMON] Dmesh service version (v1)")
 			cmd.Flags().Duration("search-common-mesh-publish-polling-duration", 0*time.Second, "[COMMON] How often does search archive poll dmesh")
 			cmd.Flags().String("search-common-action-filter-on-expr", "", "[COMMON] CEL program to whitelist actions to index. See https://github.com/dfuse-io/dfuse-eosio/blob/develop/search/README.md")
-			cmd.Flags().String("search-common-action-filter-out-expr", "", "[COMMON] CEL program to blacklist actions to index. These 2 options are used by search indexer, live and forkresolver.")
+			cmd.Flags().String("search-common-action-filter-out-expr", "account == 'eidosonecoin' || receiver == 'eidosonecoin' || (account == 'eosio.token' && (data.to == 'eidosonecoin' || data.from == 'eidosonecoin'))", "[COMMON] CEL program to blacklist actions to index. These 2 options are used by search indexer, live and forkresolver.")
 			cmd.Flags().String("search-common-dfuse-hooks-action-name", "", "[COMMON] The dfuse Hooks event action name to intercept")
 			// Router-specific flags
 			cmd.Flags().String("search-router-grpc-listen-addr", RouterServingAddr, "Address to listen for incoming gRPC requests")
@@ -659,7 +655,7 @@ to find how to install it.`)
 		Title:       "Search archive",
 		Description: "Serves historical search queries",
 		MetricsID:   "search-archive",
-		Logger:      newLoggerDef("github.com/dfuse-io/search/(archive|app/archive).*", nil),
+		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/search/(archive|app/archive).*", nil),
 		RegisterFlags: func(cmd *cobra.Command) error {
 			// These flags are scoped to search, since they are shared betwween search-router, search-live, search-archive, etc....
 			cmd.Flags().String("search-archive-grpc-listen-addr", ArchiveServingAddr, "Address to listen for incoming gRPC requests")
@@ -710,13 +706,12 @@ to find how to install it.`)
 		},
 	})
 
-	// Search Live
 	launcher.RegisterApp(&launcher.AppDef{
 		ID:          "search-live",
 		Title:       "Search live",
 		Description: "Serves live search queries",
 		MetricsID:   "search-live",
-		Logger:      newLoggerDef("github.com/dfuse-io/search/(live|app/live).*", nil),
+		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/search/(live|app/live).*", nil),
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().Uint32("search-live-tier-level", 100, "Level of the search tier")
 			cmd.Flags().String("search-live-grpc-listen-addr", LiveServingAddr, "Address to listen for incoming gRPC requests")
@@ -767,7 +762,7 @@ to find how to install it.`)
 		Title:       "Search fork resolver",
 		Description: "Search forks",
 		MetricsID:   "search-forkresolver",
-		Logger:      newLoggerDef("github.com/dfuse-io/search/(forkresolver|app/forkresolver).*", nil),
+		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/search/(forkresolver|app/forkresolver).*", nil),
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().String("search-forkresolver-grpc-listen-addr", ForkresolverServingAddr, "Address to listen for incoming gRPC requests")
 			cmd.Flags().String("search-forkresolver-http-listen-addr", ForkresolverHTTPServingAddr, "Address to listen for incoming HTTP requests")
@@ -799,13 +794,12 @@ to find how to install it.`)
 		},
 	})
 
-	// eosWS (deprecated app, scheduled to be dismantled and features migrated to dgraphql)
 	launcher.RegisterApp(&launcher.AppDef{
 		ID:          "eosws",
 		Title:       "EOSWS",
 		Description: "Serves websocket and http queries to clients",
 		MetricsID:   "eosws",
-		Logger:      newLoggerDef("github.com/dfuse-io/dfuse-eosio/eosws.*", nil),
+		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/dfuse-eosio/eosws.*", nil),
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().String("eosws-http-listen-addr", EoswsHTTPServingAddr, "Address to listen for incoming http requests")
 			cmd.Flags().Duration("eosws-graceful-shutdown-delay", time.Second*1, "delay before shutting down, after the health endpoint returns unhealthy")
@@ -858,13 +852,12 @@ to find how to install it.`)
 		},
 	})
 
-	// dGraphql
 	launcher.RegisterApp(&launcher.AppDef{
 		ID:          "dgraphql",
 		Title:       "GraphQL",
 		Description: "Serves GraphQL queries to clients",
 		MetricsID:   "dgraphql",
-		Logger:      newLoggerDef("github.com/dfuse-io/dgraphql.*", nil),
+		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/dgraphql.*", nil),
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().String("dgraphql-http-addr", DgraphqlHTTPServingAddr, "TCP Listener addr for http")
 			cmd.Flags().String("dgraphql-grpc-addr", DgraphqlGrpcServingAddr, "TCP Listener addr for gRPC")
@@ -889,6 +882,7 @@ to find how to install it.`)
 			if err != nil {
 				return nil, err
 			}
+
 			return dgraphqlEosio.NewApp(&dgraphqlEosio.Config{
 				// eos specifc configs
 				SearchAddr:    viper.GetString("dgraphql-search-addr"),
@@ -912,17 +906,16 @@ to find how to install it.`)
 		},
 	})
 
-	// eosq
 	launcher.RegisterApp(&launcher.AppDef{
 		ID:          "eosq",
 		Title:       "Eosq",
 		Description: "EOSIO Block Explorer",
 		MetricsID:   "eosq",
-		Logger:      newLoggerDef("github.com/dfuse-io/dfuse-eosio/eosq.*", nil),
+		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/dfuse-eosio/eosq.*", nil),
 		InitFunc:    nil,
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().String("eosq-http-listen-addr", EosqHTTPServingAddr, "Auth URL used to configure the dfuse js client")
-			cmd.Flags().String("eosq-api-endpoint-url", DashboardHTTPListenAddr, "API key used in eosq")
+			cmd.Flags().String("eosq-api-endpoint-url", APIProxyHTTPListenAddr, "API key used in eosq")
 			cmd.Flags().String("eosq-auth-url", JWTIssuerURL, "Auth URL used to configure the dfuse js client")
 			cmd.Flags().String("eosq-api-key", EosqAPIKey, "API key used in eosq")
 			return nil
@@ -930,7 +923,7 @@ to find how to install it.`)
 
 		FactoryFunc: func(config *launcher.BoxConfig, modules *launcher.RuntimeModules) (launcher.App, error) {
 			return eosqApp.New(&eosqApp.Config{
-				HttpListenAddr:  viper.GetString("eosq-http-listen-addr"),
+				HTTPListenAddr:  viper.GetString("eosq-http-listen-addr"),
 				APIEndpointURL:  viper.GetString("eosq-api-endpoint-url"),
 				AuthEndpointURL: viper.GetString("eosq-auth-url"),
 				ApiKey:          viper.GetString("eosq-api-key"),
@@ -938,34 +931,55 @@ to find how to install it.`)
 		},
 	})
 
-	// dashboard
 	launcher.RegisterApp(&launcher.AppDef{
 		ID:          "dashboard",
 		Title:       "Dashboard",
 		Description: "dfuse for EOSIO - dashboard",
 		MetricsID:   "dashboard",
-		Logger:      newLoggerDef("github.com/dfuse-io/dfuse-eosio/dashboard.*", nil),
+		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/dfuse-eosio/dashboard.*", nil),
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().String("dashboard-grpc-listen-addr", DashboardGrpcServingAddr, "TCP Listener addr for http")
 			cmd.Flags().String("dashboard-http-listen-addr", DashboardHTTPListenAddr, "TCP Listener addr for gRPC")
-			cmd.Flags().String("dashboard-eosws-api-addr", EoswsHTTPServingAddr, "Address of eosws api")
-			cmd.Flags().String("dashboard-dgraphql-api-addr", DgraphqlHTTPServingAddr, "Address of dgraphql api")
 			cmd.Flags().String("dashboard-eos-node-manager-api-addr", EosManagerAPIAddr, "Address of the nodeos manager api")
-			cmd.Flags().String("dashboard-mindreader-manager-api-addr", MindreaderNodeosAPIAddr, "Address of the mindreader nodeos manager api")
+			// FIXME: we can re-add when the app actually makes use of it.
+			//cmd.Flags().String("dashboard-mindreader-manager-api-addr", MindreaderNodeosAPIAddr, "Address of the mindreader nodeos manager api")
 			return nil
 		},
 		FactoryFunc: func(config *launcher.BoxConfig, modules *launcher.RuntimeModules) (launcher.App, error) {
 			return dashboard.New(&dashboard.Config{
-				GRPCListenAddr:           viper.GetString("dashboard-grpc-listen-addr"),
-				HTTPListenAddr:           viper.GetString("dashboard-http-listen-addr"),
-				EoswsHTTPServingAddr:     viper.GetString("dashboard-eosws-api-addr"),
-				DgraphqlHTTPServingAddr:  viper.GetString("dashboard-dgraphql-api-addr"),
-				EosNodeManagerAPIAddr:    viper.GetString("dashboard-eos-node-manager-api-addr"),
-				NodeosAPIHTTPServingAddr: viper.GetString("dashboard-mindreader-manager-api-addr"),
+				GRPCListenAddr:        viper.GetString("dashboard-grpc-listen-addr"),
+				HTTPListenAddr:        viper.GetString("dashboard-http-listen-addr"),
+				EosNodeManagerAPIAddr: viper.GetString("dashboard-eos-node-manager-api-addr"),
+				//NodeosAPIHTTPServingAddr: viper.GetString("dashboard-mindreader-manager-api-addr"),
 			}, &dashboard.Modules{
 				Launcher:      modules.Launcher,
 				MetricManager: modules.MetricManager,
 				DmeshClient:   modules.SearchDmeshClient,
+			}), nil
+		},
+	})
+
+	launcher.RegisterApp(&launcher.AppDef{
+		ID:          "apiproxy",
+		Title:       "API Proxy",
+		Description: "Reverse proxies all API services under one port",
+		MetricsID:   "apiproxy",
+		Logger:      launcher.NewLoggingDef("github.com/dfuse-io/dfuse-eosio/apiproxy.*", nil),
+		RegisterFlags: func(cmd *cobra.Command) error {
+			cmd.Flags().String("apiproxy-http-listen-addr", APIProxyHTTPListenAddr, "TCP Listener addr for gRPC")
+			cmd.Flags().String("apiproxy-eosws-http-addr", EoswsHTTPServingAddr, "Target address of the eosws API endpoint")
+			cmd.Flags().String("apiproxy-dgraphql-http-addr", DgraphqlHTTPServingAddr, "Target address of the dgraphql API endpoint")
+			cmd.Flags().String("apiproxy-nodeos-http-addr", NodeosAPIAddr, "Address of a queriable nodeos instance")
+			cmd.Flags().String("apiproxy-root-http-addr", EosqHTTPServingAddr, "What to serve at the root of the proxy (defaults to eosq)")
+			return nil
+		},
+		FactoryFunc: func(config *launcher.BoxConfig, modules *launcher.RuntimeModules) (launcher.App, error) {
+			return apiproxy.New(&apiproxy.Config{
+				HTTPListenAddr:   viper.GetString("apiproxy-http-listen-addr"),
+				EoswsHTTPAddr:    viper.GetString("apiproxy-eosws-http-addr"),
+				DgraphqlHTTPAddr: viper.GetString("apiproxy-dgraphql-http-addr"),
+				NodeosHTTPAddr:   viper.GetString("apiproxy-nodeos-http-addr"),
+				RootHTTPAddr:     viper.GetString("apiproxy-root-http-addr"),
 			}), nil
 		},
 	})
@@ -996,15 +1010,4 @@ func makeDirs(directories []string) error {
 	}
 
 	return nil
-}
-
-func newLoggerDef(regex string, levels []zapcore.Level) *launcher.LoggingDef {
-	if len(levels) == 0 {
-		levels = []zapcore.Level{zap.WarnLevel, zap.WarnLevel, zap.InfoLevel, zap.DebugLevel}
-	}
-
-	return &launcher.LoggingDef{
-		Levels: levels,
-		Regex:  regex,
-	}
 }